+++ src/lib/swagger-express-ts/i-swagger.ts
@@ -52,6 +52,11 @@ export interface ISwaggerOperationSchema {
     type?: string;
     items?: {$ref: string};
     $ref?: string;
+    format?: string;
+}
+
+export interface ISwaggerOperationSchemaItems {
+    $ref: string;
 }
 
 export interface ISwaggerOperationResponse {
@@ -85,9 +90,18 @@ export interface ISwaggerPath {
     delete?: ISwaggerOperation;
 }
 
+export interface ISwaggerDefinitionPropertyItems {
+    $ref: string;
+}
+
 export interface ISwaggerDefinitionProperty {
-    type: string; // Example : SwaggerDefinition.Definition.Property.Type.INTEGER
+    type?: string; // Example : SwaggerDefinition.Definition.Property.Type.INTEGER
     format?: string; // Example : SwaggerDefinition.Definition.Property.Format.INT_64
+    required?: boolean;
+    description?: string;
+    enum?: string[];
+    items?: ISwaggerDefinitionPropertyItems;
+    $ref?: string;
 }
 
 export interface ISwaggerDefinitionXML {
@@ -99,6 +113,7 @@ export interface ISwaggerDefinition {
     required?: string[];
     properties: {[key: string]: ISwaggerDefinitionProperty},
     xml?: ISwaggerDefinitionXML;
+    description?: string;
 }
 
 export interface ISwagger {