+++ src/lib/swagger-express-ts/swagger.builder.ts
@@ -6,8 +6,9 @@ import { SwaggerDefinitionConstant } from "./swagger-definition.constant";
 export interface ISwaggerBuildDefinitionModelProperty {
     /**
      * Define type of property. Example: SwaggerDefinitionConstant.Definition.Property.Type.STRING
+     * Optional.
      */
-        type: string;
+        type?: string;
 
     /**
      * Define format of property. Example: SwaggerDefinitionConstant.Definition.Property.Format.INT_64
@@ -20,9 +21,31 @@ export interface ISwaggerBuildDefinitionModelProperty {
      * Optional. Default is false.
      */
     required?: boolean;
+
+    /**
+     * Define model.
+     * Optional.
+     */
+    model?: string;
+
+    /**
+     * Define enum;
+     * Optional.
+     */
+    enum?: string [];
+
+    /**
+     * Define description.
+     */
+    description?: string;
 }
 
 export interface ISwaggerBuildDefinitionModel {
+    /**
+     * Define description.
+     */
+    description?: string;
+
     /**
      * Define all properties of model.
      */
@@ -33,13 +56,13 @@ export interface ISwaggerSecurityDefinition {
     /**
      * Define type of security.
      */
-    type: string;
+        type: string;
 
     /**
      * Define where security set.
      * Optional.
      */
-    in?: string;
+        in?: string;
 
     /**
      * Define name of security.
@@ -109,9 +132,9 @@ export interface ISwaggerBuildDefinition {
     securityDefinitions?: {[key: string]: ISwaggerSecurityDefinition};
 }
 
-export function build( buildDefinition: ISwaggerBuildDefinition ): void {
-    assert.ok( buildDefinition, "Definition are required." );
-    assert.ok( buildDefinition.info, "Informations are required. Base is { title: \"Title of my API\", version: \"1.0.0\"}" );
+export function build ( buildDefinition : ISwaggerBuildDefinition ) : void {
+    assert.ok( buildDefinition , "Definition are required." );
+    assert.ok( buildDefinition.info , "Informations are required. Base is { title: \"Title of my API\", version: \"1.0.0\"}" );
     if ( buildDefinition.basePath ) {
         SwaggerService.getInstance().setBasePath( buildDefinition.basePath );
     }
@@ -140,30 +163,7 @@ export function build( buildDefinition: ISwaggerBuildDefinition ): void {
         SwaggerService.getInstance().addSecurityDefinitions( buildDefinition.securityDefinitions );
     }
     if ( buildDefinition.models ) {
-        let definitions: {[key: string]: ISwaggerDefinition} = {};
-        for ( let modelIndex in buildDefinition.models ) {
-            let model: ISwaggerBuildDefinitionModel = buildDefinition.models[ modelIndex ];
-            let newDefinition: ISwaggerDefinition = {
-                type : SwaggerDefinitionConstant.Model.Type.OBJECT,
-                properties : {},
-                required : []
-            };
-            for ( let propertyIndex in model.properties ) {
-                let property: ISwaggerBuildDefinitionModelProperty = model.properties[ propertyIndex ];
-                let newProperty: ISwaggerDefinitionProperty = {
-                    type : property.type
-                };
-                if ( property.format ) {
-                    newProperty.format = property.format;
-                }
-                if ( property.required ) {
-                    newDefinition.required.push( propertyIndex );
-                }
-                newDefinition.properties[ propertyIndex ] = newProperty;
-            }
-            definitions[ modelIndex ] = newDefinition;
-        }
-        SwaggerService.getInstance().setDefinitions( definitions );
+        SwaggerService.getInstance().setDefinitions( buildDefinition.models );
     }
     SwaggerService.getInstance().buildSwagger();
 }