+++ CHANGELOG.md
@@ -1,3 +1,106 @@
+<a name="1.0.0-rc.1"></a>
+# [1.0.0-rc.1] (2018-04-08)
+
+## Features
+
+### Model
+
+#### Configuration
+
+example:
+
+```ts
+definition : {
+    ...
+    models : {
+        Version : {
+            properties : {
+                id : {
+                    type : SwaggerDefinitionConstant.Model.Property.Type.STRING,
+                    required : true
+                },
+                name : {
+                    type : SwaggerDefinitionConstant.Model.Property.Type.STRING,
+                    required : true
+                },
+                description : {
+                    type : SwaggerDefinitionConstant.Model.Property.Type.STRING
+                },
+                version : {
+                    type : SwaggerDefinitionConstant.Model.Property.Type.STRING
+                },
+                author: {
+                    model: "Author"
+                }
+            }
+        },
+        Author: {
+            properties: {
+                id: {
+                    type: SwaggerDefinitionConstant.Model.Property.Type.STRING,
+                    required : true
+                },
+                name : {
+                    type : SwaggerDefinitionConstant.Model.Property.Type.STRING,
+                    required : true
+                },
+            }
+        }
+    },
+    ...
+}
+```
+
+or
+
+```ts
+@ApiModel( {
+    description : "Version description" ,
+    name : "Version"
+} )
+export class VersionModel {
+
+    @ApiModelProperty( {
+        description : "Id of version" ,
+        required : true
+    } )
+    id : number;
+
+    @ApiModelProperty( {
+        description : "" ,
+        required : true
+    } )
+    name : string;
+
+    @ApiModelProperty( {
+        description : "Description of version" ,
+        required : true
+    } )
+    description : string;
+
+    @ApiModelProperty( {
+        description : "Author of version" ,
+        model : "Author"
+    } )
+    author : AuthorModel;
+}
+``` 
+
+#### Controller
+
+example:
+
+```ts
+@ApiOperationGet( {
+    ...
+    responses : {
+        200 : { description : "Success" , type : SwaggerDefinitionConstant.Response.Type.ARRAY , model : "Version" }
+    } ,
+    ...
+} )
+```
+
+
 <a name="1.0.0-beta.1"></a>
 # [1.0.0-beta.1] (2018-03-02)
 
@@ -10,23 +113,23 @@
 example:
 
 ```ts
-    app.use( swagger.express(
-        {
-            definition : {
-                ...
-                securityDefinitions : {
-                    basicAuth : {
-                        type : SwaggerDefinitionConstant.Security.Type.BASIC_AUTHENTICATION
-                    },
-                    apiKeyHeader : {
-                        type: SwaggerDefinitionConstant.Security.Type.API_KEY,
-                        in: SwaggerDefinitionConstant.Security.In.HEADER,
-                        name: "apiHeader"
-                    }
+app.use( swagger.express(
+    {
+        definition : {
+            ...
+            securityDefinitions : {
+                basicAuth : {
+                    type : SwaggerDefinitionConstant.Security.Type.BASIC_AUTHENTICATION
+                },
+                apiKeyHeader : {
+                    type: SwaggerDefinitionConstant.Security.Type.API_KEY,
+                    in: SwaggerDefinitionConstant.Security.In.HEADER,
+                    name: "apiHeader"
                 }
             }
         }
-    ) );
+    }
+) );
 ```
 
 #### Basic Authentication
@@ -34,14 +137,14 @@ example:
 Example:
 
 ```ts
+...
+@ApiOperationGet( {
     ...
-    @ApiOperationGet( {
-        ...
-        security : {
-            basicAuth : []
-        }
-    } )
-    ...
+    security : {
+        basicAuth : []
+    }
+} )
+...
 ```
 
 #### API Keys
@@ -49,14 +152,14 @@ Example:
 Example:
 
 ```ts
+...
+@ApiOperationGet( {
     ...
-    @ApiOperationGet( {
-        ...
-        security : {
-            apiKeyHeader : []
-        }
-    } )
-    ...
+    security : {
+        apiKeyHeader : []
+    }
+} )
+...
 ```
 
 ### Operations as deprecated
@@ -64,23 +167,23 @@ Example:
 Example in path:
 
 ```ts
+...
+@ApiPath( {
     ...
-    @ApiPath( {
-        ...
-        deprecated: true
-    } )
-    ...
+    deprecated: true
+} )
+...
 ```
 
 Example in operation:
 
 ```ts
+...
+@ApiOperationGet( {
     ...
-    @ApiOperationGet( {
-        ...
-        deprecated: true
-    } )
-    ...
+    deprecated: true
+} )
+...
 ```
 
 <a name="1.0.0-alpha.5"></a>